r5149: Before this gets too large, commit it.
Yet another attempt at an async winbind, not finished yet. The final idea is
that the main winbind daemon does not have any outgoing connections itself, it
asynchronously asks children, one per domain. I think with the samba3 rpc
client libs that is the best we can achieve. Probably there will be a separate
child for the idmap_ldap.

* The event loop has been reworked to allow a bit more flexible async events.
  Yes, another case of ideas stolen from Samba 4 ...

* Look at winbindd.c line 427ff. This is the heart of the client loop.

* Syncronous requests are scheduled to be send directly. The idea is that
  sooner or later no requests are sync anymore.

* For an example of an async request, see winbindd_sid.c, line 31ff.

* This uses the async API defined in winbindd_dual.c, line 248ff.

* One real bonus is the new add_trusted_domains, winbindd_util.c, line 196ff.
  Why is this special? This is an async request that has been scheduled from 
  within winbind itself without any client request. New domains are being
  scanned while normal winbind operation is continuing.

Volker