r5175: Add pdb_get_seq_num call to ldapsam.
This derives the sequence number we hand out in SAMR query_dom_info
(level 2 and 8) from the OpenLDAP 2.2 syncrepl contextCSN. This is a
perfect solution for the very common situation where posix and
samba-accounts reside in OpenLDAP as the contextCSN-timestamp is
increased on any change in the DIT. All other LDAP-Servers and
passdb-backend continue to hand out time(NULL) as sequence-number.

Due to it's dependency to OpenLDAP it must be hand-enabled (evaluating
support for syncrepl can't be automated reasonably anyway at the
moment):

        ldapsam:syncrepl_seqnum=True

In addition, when using a syncrepl consumer, one needs to set the local
replica id (rid):

        ldapsam:syncrepl_rid=[integer]

This mechanism might soon be replaced with another mechanism that Jerry
adds while adding sam-replication.

Guenther