r5444: Add adssearch.pl utility (on volkers request).
Added to samba3 due to some header and machine-account dependencies,
although it's possibly of more interest to samba4 developers.

adssearch.pl is a kind of ldapsearch + dump-filters for various
ADS-attributes. It can also register asynchronous change notifications.

./adssearch.pl -h w2k3host -D administrator@MY.REALM.NET -x -w mypass -n

and *any* change in your entire DIT will show up immediately (after a
second change of an object even as object diff). It's very interesting
to see the interaction of GPOs, the various steps of account
modification with dsa.msc, etc.

Gracefully ignore some parts of adssearch.pl that are rather immature...

Guenther