r5668: Add tests to RPC-SAMLOGON to test for user@REALM style logins.  These need a NULL domain (or a "" domain, except this breaks NTLMv2, and I need to look into it a bit more).
Add support to the Samba4 server for these logins.  This will need
extension when we handle trusted domains as a DC, as it is a principal
name, not just another format for the username.

Andrew Bartlett
(This used to be commit de02c7c222a32d2b3fb8ee8b715749b96cb647f9)