r5673: First step to get wbinfo --user-sids async. There's three calls in that: First ask the DC who holds the user for the domain groups the user is in. Then hand the list of SIDs to our own DC for alias expansion, these are the domain local groups. The last step is to expand the resulting list with the local aliases. This itself might be a blocking call with pdb_ldap for example.
Now I have to find a way to pass an arbitrary amount of data down to a child
winbindd. Have an extra_data like in the response? Or go through
winbindd_cache? Don't know, have to think about that a bit more.

BTW, sprintf_append is a little utility function to step-by-step sprintf a
string without too much copying. Might be interesting elsewere as well.

Volker