r5763: Add printerdb.
This is very, very much work in progress. Volker urgently requested to
add it regardless of it's completley unfinished and untested state. So
don't blame me :)

printerdb is planned to be a layer to make the internal nt-style
printing independent from storing all printer-relevant data in tdbs. It
currently has a tdb-backend (the must-have caching backend) and
"remote"-backends: a file-based one (for a customer project we both are
working on) and most probably an ldap-backend later on. It has also a
rough migration-tool inside net.

It compiles but really don't expect printing to work as expected.

Guenther