r6010: Change the testing order, so we test all transports for each binding option, rather than all binding options for each transport.
This means that we get to most of the tests earlier, with at least
some binding options.  (And allows us to have some confidence before
waiting for an RPC-SAMR test to finish with bigendian).

Andrew Bartlett
(This used to be commit 5c3e4df804e38037d0337e8ef288127d6cdda28a)