r6155: Fixes bug #1133
Added provision for overloading some global configuration options via the new,
per-user file ~/.smb/smb.conf.append which is read after the global config
file is read (and only if the global config file was read).  This leave the
original, BC behavior of ~/.smb/smb.conf which, if found, is read but causes
the global config file to not be read.

Also fixed a potential seg fault in to lp_dump_one().
(This used to be commit 2c5a6305bd127b1a7e65356c2b3aa5c13cd2bd74)