r6861: *finally* have a Windows client successfully importing one of my generated registry files.
The last bug was the prev and next sk record 
offsets in the sk record chain.  These should 
not be -1 but rather the first entry in the 
chain has a prev offset pointing to itself and 
the last has a next offset point to itself.

also fixed a bug where hbin blocks were not 
getting marked as dirty and therfore not getting 
flushed to disk.