r6866: A bit more cleanup work before the unixinfo pipe can be merged: sid2uid & friends are not as cleanly separated into tiny async requests as necessary. The real logic must happen in the parent. Convert sid2uid to that logic, sid2gid and the reverse functions will follow.
This stuff becomes a bit hard to follow, but somehow the complexity is there
in a completely async coding style. This is a bit like spaghetti coding with
the continuation functions being goto targets.

Tridge, I think code-structure wise this is quite close to what samba4 winbind
might end up as. Maybe this is a lesson that a user-level co-routine library
can help in terms of code clarity? Try to follow winbindd_sid2uid_async
(winbindd_async.c:1159) and tell me where all that goes...

Volker