r6911: For some weird reason, the patch I sent to Jeremy did not make it to his inbox several times.
This is an optimization for the write cache necessary for optimal performance
when sequentially writing large files.

A Win32 app writes files in chunks of, say, 256k. At the SMB layer this only
arrives as 64k chunks. This is fine for the file system algorithms that detect
that this is a growing file, if the chunks would be sent in sequence. This
however is not the case. When the XP redirector has to send a new 256k chunk
it sends one single byte at the end of the 256k chunk, probably as an
indication to NTFS that more stuff is coming and that NTFS should preallocate
stuff. This however destroys the file system prediction and *completely* kills
performance. I've seen the write system call for this one byte take 2.6
seconds where with this patch the file system happily writes 90MBytes/second
per client from several clients simultaneously. Without the patch it works
perfectly fine and correct, just *really* slow.

Volker