r6918: * fixed final issue when restoring printers to a Windows host   from a printmig.exe backup file of a Samba server.
When smbd generates a default sd for a printer, it formally 
included join the generic bits in the ACE.  Samba would always
map the specific bits in the access mask for a request for
checking against the generic bits.

Apparently, the windows spooler does not perform this
mapping of specific to generic bits internally.  Hence
this is why i was always getting access denied on the open
printer calls for the local printer object after the restore.
There really were no access bits that the spooler would look
at to give me access.

So the main change here is modify the definition 
of the PRINTER_ACE_XXX macros in rpc_spoolss.h
to include the specific bits as well as the generic 
ones.