r7997: Pointers don't kill people, people with pointers kill people...
"Honest office!  It was a mistake!  I thought the safety lock was on!"

* Fix problem setting registry values in in-memory objects

I now have printmig.exe successfully creating all of the printer
registry keys (in the tdb backend) which means that the top level
semantics are correct.
(This used to be commit 52899551070ddb8f185d53bd125ae06c192ef7b0)