r8247: remove the free of fullname in nbtname.c for now.
Metze, the ndr_token_store() code is storing temporary pointers into
this string in the token list, which means we are referring to freed
memory when we scan the token list. A better key might be a pointer
into the ndr buffer?
(This used to be commit 6a4e8cc991613773a65545eb308cf4ead75844e8)