r8890: Pull some "struct timeval" handling functions from Samba4.
Add add_timed_event, call a function at a point in the future.

Build add_idle_event on top of that, call a function regularly. This ought to
replace the similar function in module.c.

It reintroduces a call to gettimeofday per SMB, but has the potential to
simplify the timeout handling in Samba3 a lot.

Volker