r9292: More compiles fixes for autogenerated smb_interfaces ejs wrappers.
Add some code to try and work out whether a smb_interface requires a
TALLOC_CTX in the function signature.  I'm not sure whether this is
going to work as the raw api is a bit inconsistent I think.
(This used to be commit 13a101653ea7bba3b8b1f74637d981eb004484cc)