r9665: Work in progress. Rename some process_id functions.
The very first idea is to assume we have coherent fcntl locking (for example
GFS at least promises it), so we can have common messages.tdb, locking.tdb and
brlock.tdb. This might be slow as hell, but it is a start. Then we need a
notification mechanism, this right now is done by the parent smbd listening on
an UDP port, receiving USR1 signal notifications.

Volker