r9889: Pasted from mail to samba-technical after posting for review (http://samba.org/~jerry/patches/eventlog_registry_r9822_v2.patch)
The main changes here are the move to regkey_open_internal()
which allows us to open a path and get REGISTRY_KEY*.  Which
can then be used in regkey_{fetch,store}_values().
Thus giving us a common API for storing configuration
information in the virtual registry.  Since this sits
on top of the REGISTRY_OPS* layer, it means the caller
doesn't have to worry about how or where the data is
actually stored.

The major problem with this patch is the control_eventlog_hook()
now requiring that you pass in an NT_USER_TOKEN* for the
access checks necessary in regkey_open_internal().  The
only place the control_eventlog_hook() code gets invoked is
in init_registry_data(), but this poses a problem since
we don't have a valid user token then.  But I could
just create a token for root (which would be similar
to opening a key under the context of LOCAL_SYSTEM
on Windows).