r9974: Looks bigger than it is.
This is a little cleanup of the share mode API. I would like to play with
completely different schemes like the Roger Binns Revenge (tm ;-)) or other ideas
for the share mode API to get better scalability. This hides moves all of the
API to get_share_mode_lock instead of lock_share_mode. This might help to
stuff some more state in that struct share_mode_lock.

Get rid of the typedef in share_mode_entry...

Volker