r10011: This is a larger Saturday-night checkin. It compiles, but I did not run it once, this will follow tomorrow. So expect a stream of bugfix commits tomorrow.
New habits would dictate that I send it to the list before applying it, but I
see my temporary branch a bit as my playground. Sorry :-)

This is a rewrite of locking.c. It gets rid of deferred_open.tdb, this time
hopefully deadlock-free. It inserts the deferred open records into the normal
share mode entry table. This adds a bit more load on that tdb, but as the load
on that is way too high for a cluster anyway I need to find other ways to do
it even with deferred_open.tdb around.

It also adds the concept of unused share mode entries in the record. Why that?
In a cluster, the sanity check of process_exists() is very difficult to
realize. So my idea is to have a separate locking file per open file. Every
process interested in an individual share mode entry sets a read lock on the
entry. process_exists() in get_share_modes() then would see if it can get a
write lock on the entry. If the table says that the entry is still in use, but
the write lock can be acquired, then the process is obviously dead. That also
implies that share mode entries may not be moved around in the file.

Maybe for large tdb-mmap'ed share mode lists this might even be a speedup by
removing memcpy operations. Hmmmm. No. tdb_store moves stuff around anyway. We
need a tdb_fetch_lock giving a direct handle on the mmap'ed region :-).

I'll test this for use in a non-clustered Samba as well, maybe it's an
alternative implementation for huge hosts with fast directory operations to
reduce contention on locking.tdb.

Volker